% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{agg2}
\alias{agg2}
\title{Vytvori agregovane rady}
\usage{
agg2(x, upto = 8)
}
\arguments{
\item{x}{ciselny vektor}

\item{upto}{pozadovany pocet agregaci}
}
\value{
\code{list} s jednotlivymi polozkami odpovidajicimi jednotlivym agregacim se sloupci \code{DTM} (datum), \code{x} (agregovana velicina), \code{k} (uroven agregace)
}
\description{
Vytvori agregovane rady
}
\examples{
dta = get_lapv_data('AMERIKA')
agg2(dta[, .(DTM, sim_P)])
}

