% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{bil.lapv}
\alias{bil.lapv}
\title{Vytvori Bilan model s parametry ziskanymi kalibraci pro vybrane LAPV}
\usage{
bil.lapv(ID, type = "m", ...)
}
\arguments{
\item{ID}{ID lokality}

\item{type}{Typ modelu, vetsinou mesicni (tj. vychozi hodnota)}

\item{...}{ostatni parametry - predano do funkce \code{bil.new}}
}
\value{
bilan model objekt
}
\description{
Vytvori Bilan model s parametry ziskanymi kalibraci pro vybrane LAPV
}
\examples{
dta  = get_lapv_data("AMERIKA")
b = bil.lapv("AMERIKA")

bil.set.values(b, dta[, .(DTM, P = obs_P, T = obs_T)])
bil.pet(b)
res = bil.run(b)
}

