% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{correct}
\alias{correct}
\title{Proved korekci systematickych chyb pomoci kvantilove metody}
\usage{
correct(dta, ...)
}
\arguments{
\item{dta}{data pro LAPV - typicky vysledek volani \code{get_lapv_data}}

\item{...}{ostatni parametry - predane do funkce \code{loess} slouzici k vyhlazeni odchylek}
}
\value{
data s korigovanymi sloupci (sloupce \code{cor_P} a \code{cor_T})
}
\description{
Proved korekci systematickych chyb pomoci kvantilove metody
}
\examples{
dta = get_lapv_data("AMERIKA")
cdta = correct(dta)
}

