% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{get_lapv_data}
\alias{get_lapv_data}
\title{Vybere data pro jednu LAPV}
\usage{
get_lapv_data(ID)
}
\arguments{
\item{ID}{identifikator lokality}
}
\value{
\code{data.table} s daty pro zvolenou LAPV
}
\description{
Vybere data pro jednu LAPV
}
\examples{
get_lapv_data('AMERIKA')
}

