% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{hurst}
\alias{hurst}
\title{Spocíta statistiky dle Hursta}
\usage{
hurst(x, N = 10:100)
}
\arguments{
\item{x}{vektor pro vypocet statistik}

\item{N}{vektor resamplovanych delek}
}
\value{
data.frame s vysledky
}
\description{
Spocíta statistiky dle Hursta
}

