% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{mon2yr}
\alias{mon2yr}
\title{Agreguj na roky}
\usage{
mon2yr(dta)
}
\arguments{
\item{dta}{data pro LAPV}
}
\value{
\code{data.table} s agregovanym datasetem
}
\description{
Agreguj na roky
}
\examples{
dta = get_lapv_data('AMERIKA')
mon2yr(dta)
}

