% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{toyGen}
\alias{toyGen}
\title{Generator nahodnych velicin}
\usage{
toyGen(x, n = length(x), proces = "WN", seed = NULL, burnin = 100)
}
\arguments{
\item{x}{ciselny vektor - generovana data maji stejny prumer a smerodatnou odchylku jako tato velicina}

\item{n}{delka generovaneho vektoru, vychozi \code{n = length(x)}}

\item{proces}{specifikace procesu generujiciho nahodnou velicinu - jedno z \code{WN} - bily sum, \code{ARi} - autoregresni proces, kde \code{i} je rad procesu (tedy napr. \code{AR1}, \code{AR2}, atd.), \code{FGN} - fractional Gaussian noise}

\item{seed}{seed nahodneho generatoru}
}
\value{
vygenerovany vektor
}
\description{
Generator nahodnych velicin
}
\examples{
dta = get_lapv_data('AMERIKA')
x = dta[, .(obs_RM)]
toyGen(x, proces = 'WN')
toyGen(x, proces = 'AR2')
toyGen(x, proces = 'FGN')
}

