% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{agg2}
\alias{agg2}
\title{Vytvoří agregované řady}
\usage{
agg2(x, upto = 8)
}
\arguments{
\item{x}{číselný vektor}

\item{upto}{požadovaný počet agregací}
}
\value{
\code{list} s jednotlivými položkami odpovídajícími jednotlivým agregacím se sloupci \code{DTM} (datum), \code{x} (agregovaná veličina), \code{k} (úroveň agregace)
}
\description{
Vytvoří agregované řady
}
\examples{
dta = get_lapv_data('AMERIKA')
agg2(dta[, .(DTM, sim_P)])
}

