% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{hurst}
\alias{hurst}
\title{Spočítá statistiky dle Hursta}
\usage{
hurst(x, N = 10:100)
}
\arguments{
\item{x}{vektor pro výpočet statistik}

\item{N}{vektor resamplovaných délek}
}
\value{
data.frame s výsledky
}
\description{
Spočítá statistiky dle Hursta
}

