% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{plotGen}
\alias{plotGen}
\title{Vykresli casovou radu a kumulativni sumu pro generovane rady}
\usage{
plotGen(x, seed = NULL, ylim = c(-3000, 3000))
}
\arguments{
\item{x}{vzro pro generatory}

\item{seed}{seed pro generatory}

\item{ylim}{rozsah}
}
\value{
nic
}
\description{
Vykresli casovou radu a kumulativni sumu pro generovane rady
}

