% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapv.R
\name{toyGen}
\alias{toyGen}
\title{Generátor náhodných veličin}
\usage{
toyGen(x, n = length(x), proces = "WN", seed = NULL, burnin = 100)
}
\arguments{
\item{x}{číselný vektor - generovaná data mají stejný průměr a směrodatnou odchylku jako tato veličina}

\item{n}{délka generovaného vektoru, výchozí \code{n = length(x)}}

\item{proces}{specifikace procesu generující náhodnou veličinu - jedno z \code{WN} - bílý šum, \code{ARi} - autoregresní proces, kde \code{i} je řád procesu (tedy např. \code{AR1}, \code{AR2}, atd.), \code{FGN} - fractional Gaussian noise}

\item{seed}{seed náhodného generátoru}
}
\value{
vygenerovaný vektor
}
\description{
Generátor náhodných veličin
}
\examples{
dta = get_lapv_data('AMERIKA')
x = dta[, .(obs_RM)]
toyGen(x, proces = 'WN')
toyGen(x, proces = 'AR2')
toyGen(x, proces = 'FGN')
}

